% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data documentation.R
\docType{data}
\name{davos}
\alias{davos}
\title{Davos dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19758 rows and 15 columns.}
\usage{
data(davos)
}
\description{
Lists of participants of the World Economic Forum from 2013-2019 period.
}
\details{
The data was originally published by Quartz and has been assembled and cleaned with a focus on the organizations.
}
\keyword{datasets}
